h3

Module quic

QUIC Transport traits

This module includes traits and types meant to allow being generic over any QUIC implementation.

Structs

InvalidStreamId

Invalid StreamId, for example because it’s too large

StreamId

Identifier for a stream

WriteBuf

Wrap frames to encode their header on the stack before sending them on the wire

Enums

ConnectionErrorIncoming

Error type to communicate that the quic connection was closed

StreamErrorIncoming

Error type to communicate that the stream was closed

Traits

BidiStream

Optional trait to allow “splitting” a bidirectional stream into two sides.

Connection

Trait representing a QUIC connection.

OpenStreams

Trait for opening outgoing streams

RecvStream

A trait describing the “receive” actions of a QUIC stream.

SendStream

A trait describing the “send” actions of a QUIC stream.

SendStreamUnframed

Allows sending unframed pure bytes to a stream. Similar to AsyncWrite