h3::quic

Trait OpenStreams

pub trait OpenStreams<B: Buf> {
    type BidiStream: SendStream<B> + RecvStream;
    type SendStream: SendStream<B>;

    // Required methods
    fn poll_open_bidi(
        &mut self,
        cx: &mut Context<'_>,
    ) -> Poll<Result<Self::BidiStream, StreamErrorIncoming>>;
    fn poll_open_send(
        &mut self,
        cx: &mut Context<'_>,
    ) -> Poll<Result<Self::SendStream, StreamErrorIncoming>>;
    fn close(&mut self, code: Code, reason: &[u8]);
}

Trait for opening outgoing streams

Required Associated Types

type BidiStream: SendStream<B> + RecvStream

The type produced by poll_open_bidi()

type SendStream: SendStream<B>

The type produced by poll_open_send()

Required Methods

fn poll_open_bidi( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<Self::BidiStream, StreamErrorIncoming>>

Poll the connection to create a new bidirectional stream.

fn poll_open_send( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<Self::SendStream, StreamErrorIncoming>>

Poll the connection to create a new unidirectional stream.

fn close(&mut self, code: Code, reason: &[u8])

Close the connection immediately