h3::server

Struct Builder

pub struct Builder { /* private fields */ }

Builder of HTTP/3 server connections.

Implementations

impl Builder

pub fn max_field_section_size(&mut self, value: u64) -> &mut Self

Set the maximum header size this client is willing to accept

See header size constraints section of the specification for details.

pub fn send_grease(&mut self, value: bool) -> &mut Self

Send grease values to the Client. See setting, frame and stream for more information.

pub fn enable_webtransport(&mut self, value: bool) -> &mut Self

Indicates to the peer that WebTransport is supported.

See: establishing a webtransport session

Server: Supporting for webtransport also requires setting enable_extended_connect enable_datagram and max_webtransport_sessions.

pub fn enable_extended_connect(&mut self, value: bool) -> &mut Self

Enables the extended CONNECT protocol required for various HTTP/3 extensions.

pub fn max_webtransport_sessions(&mut self, value: u64) -> &mut Self

Limits the maximum number of WebTransport sessions

pub fn enable_datagram(&mut self, value: bool) -> &mut Self

Indicates that the client or server supports HTTP/3 datagrams

See: https://www.rfc-editor.org/rfc/rfc9297#section-2.1.1

impl Builder

pub async fn build<C, B>( &self, conn: C, ) -> Result<Connection<C, B>, ConnectionError>
where C: Connection<B>, B: Buf,

Build an HTTP/3 connection from a QUIC connection

This method creates a Connection instance with the settings in the Builder.