h3::error

Enum StreamError

#[non_exhaustive]
pub enum StreamError {

    #[non_exhaustive]
    StreamError {
        code: Code,
        reason: String,
    },

    #[non_exhaustive]
    RemoteTerminate {
        code: Code,
    },

    #[non_exhaustive]
    ConnectionError(ConnectionError),

    #[non_exhaustive]
    HeaderTooBig {
        actual_size: u64,
        max_size: u64,
    },

    #[non_exhaustive]
    RemoteClosing,

    #[non_exhaustive]
    Undefined(Box<dyn Error + Send + Sync>),
}

This enum represents a stream error

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

#[non_exhaustive]
StreamError

The error occurred on the stream

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

code: Code

The error code

reason: String

The error reason

#[non_exhaustive]
RemoteTerminate

The remote peer terminated the corresponding stream side

Either Reset on peers sending side or StopSending on peers receiving side

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

code: Code

Reset code received from the peer

#[non_exhaustive]
ConnectionError(ConnectionError)

The error occurred on the connection

#[non_exhaustive]
HeaderTooBig

Error is used when violating the MAX_FIELD_SECTION_SIZE

This can mean different things depending on the context When sending a request, this means, that the request cannot be sent because the header is larger then permitted by the server When receiving a request, this means, that the server sent a

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

actual_size: u64

The actual size of the header block

max_size: u64

The maximum size of the header block

#[non_exhaustive]
RemoteClosing

Received a GoAway frame from the remote

Stream operations cannot be performed

#[non_exhaustive]
Undefined(Box<dyn Error + Send + Sync>)

Undefined error propagated by the quic layer

Implementations

impl StreamError

pub fn is_h3_no_error(&self) -> bool

Returns if the error is H3_NO_ERROR