h3::error

Struct Code

pub struct Code { /* private fields */ }

An HTTP/3 โapplication error codeโ.

Implementations

impl Code

pub const fn value(&self) -> u64

Numerical error code

See https://www.rfc-editor.org/rfc/rfc9114.html#errors and https://www.rfc-editor.org/rfc/rfc9000.html#error-codes

impl Code

pub const H3_DATAGRAM_ERROR: Code

Datagram or capsule parse error See: https://www.rfc-editor.org/rfc/rfc9297#section-5.2

pub const H3_NO_ERROR: Code

No error. This is used when the connection or stream needs to be closed, but there is no error to signal.

pub const H3_GENERAL_PROTOCOL_ERROR: Code

Peer violated protocol requirements in a way that does not match a more specific error code, or endpoint declines to use the more specific error code.

pub const H3_INTERNAL_ERROR: Code

An internal error has occurred in the HTTP stack.

pub const H3_STREAM_CREATION_ERROR: Code

The endpoint detected that its peer created a stream that it will not accept.

pub const H3_CLOSED_CRITICAL_STREAM: Code

A stream required by the HTTP/3 connection was closed or reset.

pub const H3_FRAME_UNEXPECTED: Code

A frame was received that was not permitted in the current state or on the current stream.

pub const H3_FRAME_ERROR: Code

A frame that fails to satisfy layout requirements or with an invalid size was received.

pub const H3_EXCESSIVE_LOAD: Code

The endpoint detected that its peer is exhibiting a behavior that might be generating excessive load.

pub const H3_ID_ERROR: Code

A Stream ID or Push ID was used incorrectly, such as exceeding a limit, reducing a limit, or being reused.

pub const H3_SETTINGS_ERROR: Code

An endpoint detected an error in the payload of a SETTINGS frame.

pub const H3_MISSING_SETTINGS: Code

No SETTINGS frame was received at the beginning of the control stream.

pub const H3_REQUEST_REJECTED: Code

A server rejected a request without performing any application processing.

pub const H3_REQUEST_CANCELLED: Code

The request or its response (including pushed response) is cancelled.

pub const H3_REQUEST_INCOMPLETE: Code

The clientโs stream terminated without containing a fully-formed request.

pub const H3_MESSAGE_ERROR: Code

An HTTP message was malformed and cannot be processed.

pub const H3_CONNECT_ERROR: Code

The TCP connection established in response to a CONNECT request was reset or abnormally closed.

pub const H3_VERSION_FALLBACK: Code

The requested operation cannot be served over HTTP/3. The peer should retry over HTTP/1.1.

pub const QPACK_DECOMPRESSION_FAILED: Code

The decoder failed to interpret an encoded field section and is not able to continue decoding that field section.

pub const QPACK_ENCODER_STREAM_ERROR: Code

The decoder failed to interpret an encoder instruction received on the encoder stream.

pub const QPACK_DECODER_STREAM_ERROR: Code

The encoder failed to interpret a decoder instruction received on the decoder stream.