h3
Module quic

QUIC Transport traits

This module includes traits and types meant to allow being generic over any QUIC implementation.

Structs

InvalidStreamId

Invalid StreamId, for example because it’s too large

StreamId

Identifier for a stream

WriteBuf

Wrap frames to encode their header on the stack before sending them on the wire

Traits

BidiStream

Optional trait to allow “splitting” a bidirectional stream into two sides.

Connection

Trait representing a QUIC connection.

Error

Trait that represent an error from the transport layer

OpenStreams

Trait for opening outgoing streams

RecvDatagramExt

Extends the Connection trait for receiving datagrams

RecvStream

A trait describing the “receive” actions of a QUIC stream.

SendDatagramExt

Extends the Connection trait for sending datagrams

SendStream

A trait describing the “send” actions of a QUIC stream.

SendStreamUnframed

Allows sending unframed pure bytes to a stream. Similar to AsyncWrite