h3::quic
Trait OpenStreams

pub trait OpenStreams<B: Buf> {
    type BidiStream: SendStream<B> + RecvStream;
    type SendStream: SendStream<B>;
    type RecvStream: RecvStream;
    type Error: Into<Box<dyn Error>>;

    // Required methods
    fn poll_open_bidi(
        &mut self,
        cx: &mut Context<'_>,
    ) -> Poll<Result<Self::BidiStream, Self::Error>>;
    fn poll_open_send(
        &mut self,
        cx: &mut Context<'_>,
    ) -> Poll<Result<Self::SendStream, Self::Error>>;
    fn close(&mut self, code: Code, reason: &[u8]);
}

Trait for opening outgoing streams

Required Associated Types

type BidiStream: SendStream<B> + RecvStream

The type produced by poll_open_bidi()

type SendStream: SendStream<B>

The type produced by poll_open_send()

type RecvStream: RecvStream

The type of the receiving part of BidiStream

type Error: Into<Box<dyn Error>>

Error type yielded by these trait methods

Required Methods

fn poll_open_bidi( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<Self::BidiStream, Self::Error>>

Poll the connection to create a new bidirectional stream.

fn poll_open_send( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<Self::SendStream, Self::Error>>

Poll the connection to create a new unidirectional stream.

fn close(&mut self, code: Code, reason: &[u8])

Close the connection immediately