h3::quic
Trait RecvStream

pub trait RecvStream {
    type Buf: Buf;
    type Error: Into<Box<dyn Error>>;

    // Required methods
    fn poll_data(
        &mut self,
        cx: &mut Context<'_>,
    ) -> Poll<Result<Option<Self::Buf>, Self::Error>>;
    fn stop_sending(&mut self, error_code: u64);
    fn recv_id(&self) -> StreamId;
}

A trait describing the “receive” actions of a QUIC stream.

Required Associated Types

type Buf: Buf

The type of Buf for data received on this stream.

type Error: Into<Box<dyn Error>>

The error type that can occur when receiving data.

Required Methods

fn poll_data( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<Option<Self::Buf>, Self::Error>>

Poll the stream for more data.

When the receive side will no longer receive more data (such as because the peer closed their sending side), this should return None.

fn stop_sending(&mut self, error_code: u64)

Send a STOP_SENDING QUIC code.

fn recv_id(&self) -> StreamId

Get QUIC send stream id